refactor: reorganize the sqlglot scalar compiler layout - part 2
